Improve write operation spec. For #4 and #11.
Mainly modify the spec on how to do upsert via update type.

It's straightfoward for Mongo upsert.

We translate Postgres upsert to `INSERT INTO ... ON CONFLICT DO ...`
statement. Treat `query` as the conflict target, if any conflict
do update with `write`, otherwise insert the merge result of
`query` and `write` if no conflict.